refactor: extract AppLifecycleDelegate.swift from BugNarratorApp.swift
Extracts the AppLifecycleDelegate class (NSApplicationDelegate, 15 lines)
from BugNarratorApp.swift into its own file.

Byte-identical move; SHA-256 b2b550cbe7492e7cc145d8728d67c24e8f8b5bdc279c3491ed8f7e64710d5a86
verified against origin/main:BugNarratorApp.swift lines 4-18.

Imports: AppKit (for NSApplicationDelegate).

Reduces BugNarratorApp.swift from 233 to 216 lines.

Tests: 667 (unchanged).
